requirements.txt
flask
sqlalchemy
flask-sqlalchemy
pymysql
psycopg2-binary
python-dotenv
gunicorn



